Fix the "Use Item" functionality.
When you use an item, a cooldown time starts where you have to wait until you can use an item again. The reasoning for this is because otherwise, FixedUpdate runs so often that even if you pushed down a "Use Item" key for a split second, FixedUpdate still utilized multiple items.